Fix push notifications on Windows
Distinguish foreground/background notifications like on other
platforms.

Extract launch parameters as additionalData. This mimics what is
possible out of the box in iOS and Android for the common case of
receiving a key-value dictionary.

Allow silencing OS foreground notifications. This is intended to emulate
the behavior of iOS/Android and let the app handle foreground
notifications by itself.

Co-authored-by: Stefano Motta <stefano.motta@webratio.com>
Co-authored-by: Lorenzo Maghini <lorenzo.maghini@webratio.com>
WR-Issue: #12445